reports to new module system
