--- modules/reports/index.php
@@ -1,4 +1,0 @@
-<?php
-$smarty -> assign('pageActive', 'report');
-$smarty -> assign('active_tab', '#home');
-?>
